Fix the showing of occurred in the listing of logs
There was special logic to do special stuff for a particular version of
metadata.  The logic should have been reversed.  This manifested a
second issue where the version value had whitespce (I could have
probably done an `in` instead) so this was changed to strip.  I also
just took care of whitespace issues with tags too.